+++ src/sim/Sim.tsx
@@ -1,165 +1,8 @@
 import React from 'react';
 import fp from 'lodash/fp';
 import { KdTreeMap } from '@thi.ng/geom-accel';
-import { CanvasRender } from './render';
-
-// Helper type for a maybe-define for thing types
-export type ThingHas<T> = Partial<Record<ThingType, T>>;
-
-const getRes = (thing: Thing, resType: ThingType) => thing.resources[resType] ?? 0;
-const mathRes = (thing: Thing, resType: ThingType, value: number) => {
-	const newValue = getRes(thing, resType) + value;
-	thing.resources[resType] = newValue;
-	return newValue;
-};
-
-/**
- * Our simulation has a tick (current time) and things
- *
- * Things are organized into a 2D k-d tree (for now) with a collection of things to allow overlapping points
- *
- * There's a graph of connections between things as well for resource sharing / related components
-*/
-export type SimState = {
-	things: KdTreeMap<number[], Thing[]>,
-	tick: number
-};
-
-
-/**
- * A thing is something, somewhere, with some resources (which are counts of things)
- */
-export type Thing = {
-	type: ThingType,
-	pos: number[],
-	resources: ThingHas<number>,
-};
-
-export enum ThingType {
-	Water = 'Water',
-	Seed = 'Seed',
-	Root = 'Root',
-	Cloud = 'Cloud',
-	RainCloud = 'RainCloud',
-};
-
-// Helpers for working with the stacked kd map
-const addThing = (thing: Thing, things: SimState['things']) => {
-	const existing = things.get(thing.pos);
-	if (existing !== undefined) {
-		existing.push(thing);
-	} else {
-		things.set(thing.pos, [thing]);
-	}
-	return things;
-};
-const removeThing = (thing: Thing, things: SimState['things']) => {
-	const remaining = (things.get(thing.pos) ?? [])
-		.filter(x => x !== thing);
-	if (remaining.length) {
-		things.set(thing.pos, remaining);
-	} else {
-		things.remove(thing.pos);
-	}
-	return things;
-};
-
-
-export type Control = (self: Thing, state: SimState, delta: number) => void;
-
-// -------- Controls
-const absorbWater: Control = (self: Thing, state: SimState) => {
-	const DISTANCE = 2;
-	const MAX_CELLS = (DISTANCE - 1) * 9;
-	const nearestWater = state.things.queryValues(self.pos, 2, MAX_CELLS)
-		.flat()
-		.find(x => x.type === ThingType.Water);
-
-	if (nearestWater) {
-		removeThing(nearestWater, state.things);
-		mathRes(self, ThingType.Water, 1);
-	}
-};
-
-const rootDown: Control = (self: Thing, state: SimState) => {
-	if (getRes(self, ThingType.Water) > 2) {
-		const newRoot = root();
-		newRoot.pos[0] = self.pos[0];
-		newRoot.pos[1] = self.pos[1] + 1; // down 1
-		addThing(newRoot, state.things);
-		mathRes(self, ThingType.Water, -2);
-	}
-};
-
-const condenseWater: Control = (self: Thing, state: SimState) => {
-	// Gather water
-	if (state.tick % 3 === 0) {
-		mathRes(self, ThingType.Water, 5);
-	}
-
-
-	// Become a rain cloud
-	if (getRes(self, ThingType.Water) > 10) {
-		self.type = ThingType.RainCloud;
-	}
-};
-
-const rain: Control = (self: Thing, state: SimState) => {
-	let maxDrops = 4;
-	while (maxDrops > 0 && getRes(self, ThingType.Water) > 0) {
-		const newWater = water();
-		newWater.pos[0] = self.pos[0];
-		newWater.pos[1] = self.pos[1];
-		addThing(newWater, state.things);
-		mathRes(self, ThingType.Water, -1);
-		maxDrops--;
-	}
-
-	// Become a cloud again
-	if (self.resources[ThingType.Water] === 0) {
-		self.type = ThingType.Cloud;
-	}
-};
-
-export const FixedControls: ThingHas<Control[]> = {
-	[ThingType.Seed]: [absorbWater, rootDown],
-	[ThingType.Root]: [absorbWater],
-	[ThingType.Cloud]: [condenseWater],
-	[ThingType.RainCloud]: [rain],
-};
-
-// --------------------- Things
-const baseThing = (pos: number[]) => ({
-	resources: {},
-	pos,
-});
-
-const water = (pos: number[] = [0, 0]) => ({
-	...baseThing(pos),
-	type: ThingType.Water,
-});
-
-const seed = (pos: number[] = [0, 0]) => ({
-	...baseThing(pos),
-	type: ThingType.Seed,
-	resources: {
-		[ThingType.Water]: 0,
-	},
-});
-
-const root = (pos: number[] = [0, 0]) => ({
-	...baseThing(pos),
-	type: ThingType.Root,
-});
-
-const cloud = (pos: number[] = [0, 0]) => ({
-	...baseThing(pos),
-	type: ThingType.Cloud,
-	resource: {
-		[ThingType.Water]: 0,
-	}
-});
-
+import { CanvasRender, SvgRender } from './render';
+import { addThing, cloud, FixedControls, seed, SimState, Thing, water } from './things';
 
 // -- Sim
 export type SerializedSimState = string;
@@ -202,7 +45,7 @@ const emptyState = () => ({
 		// A seed
 		seed(),
 		// A cloud
-		cloud(),
+		cloud([0, -1]),
 	].reduce((thingMap, thing) => addThing(thing, thingMap), new KdTreeMap<number[], Thing[]>(2)),
 	tick: 0,
 });
@@ -288,7 +131,8 @@ export class Sim extends React.Component<SimProps, SimEltState> {
 		return (
 			<div className="sim-container">
 				<div className="sim-render-area">
-					<CanvasRender state={this.state.sim} onResize={dims => this.setCanvasDims(dims)} />
+					{/* <CanvasRender state={this.state.sim} onResize={dims => this.setCanvasDims(dims)} /> */}
+					<SvgRender state={this.state.sim} onResize={dims => this.setCanvasDims(dims)} />
 					{/* {[...this.state.things.values()].flatMap((things, i) => (
 						things.map((thing, j) => (
 							<DisplayThing thing={thing} key={`${i},${j}`} />
